Switch tests to use `assert` library

+++ task.go
@@ -6,7 +6,7 @@ import (
 )
 
 type Task struct {
-	ID                int64     `json:"id"`
+	ID                int       `json:"id"`
 	Name              string    `json:"name"`
 	BillableByDefault bool      `json:"billable_by_default"`
 	DefaultHourlyRate float64   `json:"default_hourly_rate"`
